Restructured project
Moved all actual entities to entity module, which exports public facing
ones
