+++ src/game/engine/entity/__init__.py
@@ -1,0 +1,7 @@
+from .player import Player
+from .platform import Platform, Axis
+from .game_object import GameObject
+from .physics_object import PhysicsObject
+from .star import Star
+from .obstacle import Obstacle
+from .enemy import Enemy
