+++ src/game/engine/entity/obstacle.py
