+++ src/game/engine/entity/star.py
